Move kubefed2/federate/enable to kubefed2/enable pkg
