+++ pkg/kubefed2/enable/schema.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package federate
+package enable
 
 import (
 	"fmt"
