+++ pkg/kubefed2/enable/util.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package federate
+package enable
 
 import (
 	"fmt"
