+++ pkg/kubefed2/kubefed2.go
@@ -25,6 +25,7 @@ import (
 	apiserverflag "k8s.io/apiserver/pkg/util/flag"
 	"k8s.io/client-go/tools/clientcmd"
 
+	"github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/enable"
 	"github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/federate"
 	"github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/util"
 )
@@ -53,7 +54,7 @@ func NewKubeFed2Command(out io.Writer) *cobra.Command {
 	flag.CommandLine.Parse(nil)
 
 	fedConfig := util.NewFedConfig(clientcmd.NewDefaultPathOptions())
-	rootCmd.AddCommand(federate.NewCmdTypeEnable(out, fedConfig))
+	rootCmd.AddCommand(enable.NewCmdTypeEnable(out, fedConfig))
 	rootCmd.AddCommand(federate.NewCmdTypeDisable(out, fedConfig))
 	rootCmd.AddCommand(NewCmdJoin(out, fedConfig))
 	rootCmd.AddCommand(NewCmdUnjoin(out, fedConfig))
